requirements.txt
azure-identity
openai
python-dotenv
pydantic
rich
azure-ai-inference
dotenv-azd
aiohttp
faker
autogen-agentchat
autogen-ext[openai]
azure-ai-inference==1.0.0b9
openai-agents
semantic-kernel
langgraph
langchain==1.0.0a3
langchain_openai
pydantic-ai==1.0.0b1
llama-index
llama-index-llms-azure-openai
llama-index-llms-openai-like
llama-index-embeddings-azure-openai
smolagents
langchain-mcp-adapters
agent-framework @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/main
agent-framework-azure @ git+https://github.com/microsoft/agent-framework.git@main#subdirectory=python/packages/azure
